requirements.txt
langchain
langgraph
langchain-community
langchain-openai
langchain-mcp-adapters
psycopg2-binary
matplotlib
